DRY archivable in a concern
