+++ spec/models/diagnosis_spec.rb
@@ -13,19 +13,6 @@ RSpec.describe Diagnosis, type: :model do
     is_expected.to validate_inclusion_of(:step).in_array(Diagnosis::AUTHORIZED_STEPS)
   end
 
-  describe 'archive' do
-    let(:diagnosis) { create :diagnosis }
-
-    before do
-      diagnosis.archive!
-    end
-
-    it('archives the diagnosis') do
-      expect(Diagnosis.all.count).to eq 1
-      expect(Diagnosis.not_archived.count).to eq 0
-    end
-  end
-
   describe 'scopes' do
     describe 'in progress' do
       subject { Diagnosis.in_progress.count }
